api/requirements.txt
txt
whitenoise==6.7.0
tortoise-orm==0.21.4
fastapi==0.111.0
rq==1.16.2
asyncpg
fastapi_cors==0.0.6
pyjwt
passlib==1.7.4
bcrypt==4.2.0
aerich==0.7.1 # do NOT update to 0.7.2
requests==2.32.3
aiohttp==3.9.5
itsdangerous==2.2.0
mailjet_rest
anthropic==0.31.0
openai==1.35.14
stripe
